Store Flask-specific configuration separately
Rather than setting all of the values in the Flask configuration
manually, we can use the handy `app.config.from_object` method as long
as such a configuration object exists. We cannot configure from
`config.py` because there are other non-Flask-related configuration
values present there and I want to avoid the potential for namespace
collisions now and in the future. Therefore I refactored the
Flask-specific configuration values in `config.py` into the FlaskConfig
object, and the configuration is loaded all at once after creating the
app, instead of piecemeal in different places throughout the application
code.

This is a prerequisite for cleanly implementing server-side sessions
via Flask-Session, because it has a lot of configuration options.