Server-side sessions with Flask-Session
Replaces the default Flask session cookies with server side
sessions. The cookies are still signed, but now only contain a UUID that
refers to an entry in local storage, which avoids leaking the session
contents to the client.

Since the session contains the random challenge string, preventing
clients from directly accessing the session contents is critical to the
security of this system.